+++ public/data/FeedsData.json
@@ -1,0 +1,272 @@
+[
+  {
+    "feedId": 42,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong0.jpg",
+    "content": "42모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 43,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong1.jpg",
+    "content": "43모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 44,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong2.jpg",
+    "content": "44모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 45,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong.jpg",
+    "content": "4545모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 46,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong3.jpg",
+    "content": "4646모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 47,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong4.jpg",
+    "content": "4747모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 48,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong5.jpg",
+    "content": "4848모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 49,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong6.jpg",
+    "content": "4949모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 50,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong7.jpg",
+    "content": "5050모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 51,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong8.jpg",
+    "content": "5151모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 52,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong9.jpg",
+    "content": "5252모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 52,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong10.jpg",
+    "content": "5252모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 52,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong11.jpg",
+    "content": "5252모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 52,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong12.jpg",
+    "content": "5252모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  },
+  {
+    "feedId": 52,
+    "authorId": 1,
+    "nickName": "heywoo",
+    "profileImage": null,
+    "updatedAt": "2024-01-25T08:43:33.085Z",
+    "lowTemperature": -5,
+    "highTemperature": 3,
+    "imageUrl": "/weather/images/TestImg/dalkong13.jpg",
+    "content": "5252모두 감기 조심하세요ㅠㅠ #춥다 #감기조심",
+    "feedTags": "감기조심,춥다",
+    "weatherCondition": "snow",
+    "weatherConditionImage": "",
+    "likeCount": "0",
+    "commentCount": "0",
+    "isLiked": "0",
+    "isBookmarked": "0"
+  }
+]
