+++ src/pages/Main/Components/FeedSection/FeedSection.scss
@@ -6,5 +6,6 @@
   flex-direction: row;
   background-color: $tertiaryColor;
   flex-grow: 3;
+  padding: 2px;
 }
