+++ src/pages/Main/Components/ImageTile/ImageTile.jsx
@@ -1,0 +1,9 @@
+
+const ImageTile = ({imageUrl}) => {
+
+  return <div className="imageTile">
+    <img src={imageUrl} />
+  </div>
+}
+
+export default ImageTile;
