+++ src/pages/Main/Components/ImageTile/ImageTile.scss
@@ -1,0 +1,10 @@
+
+
+.imageTile {
+  width: 100%;
+  aspect-ratio: 1 / 1;
+  img {
+    object-fit: cover;
+  }
+}
+
