+++ src/pages/Main/Components/WeatherSection/components/Icon/Icon.scss
@@ -1,0 +1,15 @@
+@charset 'utf-8';
+@import '../../../../../../styles/partials/mixin.scss', '../../../../../../styles/partials/variables';
+
+.iconWrapper {
+  @include FlexCenter;
+  width: 100%;
+  height: 100%;
+  
+  svg {
+    width: 100%;
+    height: 100%;
+  }
+}
+
+
